add file parser function
