+++ .github/workflows/run-tests.yml
@@ -1,26 +1,20 @@
-name: Node.js CI
+name: Run Tests
 
 on:
-  push:
-    branches: [ $default-branch ]
   pull_request:
 
 jobs:
-  build:
-
+  test:
     runs-on: ubuntu-latest
-
     strategy:
       matrix:
         node-version: [12.x, 15.x]
-
     steps:
     - uses: actions/checkout@v2
     - name: Use Node.js ${{ matrix.node-version }}
       uses: actions/setup-node@v1
       with:
         node-version: ${{ matrix.node-version }}
-    - run: npm ci
-    - run: npm run build --if-present
     - run: npm test
+    - run: npm run parse
