+++ package.json
@@ -2,12 +2,12 @@
   "name": "robot-test",
   "version": "1.0.0",
   "description": "",
-  "main": "src/index.js",
-  "type": "module",
+  "main": "src/index.ts",
   "engine": {
     "node": ">=12.18.3"
   },
   "scripts": {
+    "parse": "ts-node src/index.ts",
     "test": "jest",
     "coverage": "jest --coverage"
   },
@@ -19,6 +19,7 @@
   "devDependencies": {
     "@types/jest": "^26.0.22",
     "jest": "^26.6.3",
-    "ts-jest": "^26.5.4"
+    "ts-jest": "^26.5.4",
+    "ts-node": "^9.1.1"
   }
 }
