+++ test-data/test1.txt
@@ -1,0 +1,2 @@
+N 0 0
+M1RM4L3M2
